Introduce HA support
Change the client behaviour in such a way that in case
a second metalbond server distributing the exact some
routes as the "main" server is added to the server list
then the client is aware of the second server when adding
removing the routes from its routing table hence withdraws
already received routes only if both servers withdraw
the routes in question.

Single server operation is same as before. (Backward compatible)

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>